Refactor GetSecurityGroupsResponse; Remove code duplicated.
