+++ src/main/java/cloud/fogbow/ras/core/plugins/interoperability/openstack/OpenStackCloudUtils.java
@@ -5,7 +5,8 @@ import cloud.fogbow.common.exceptions.InvalidParameterException;
 import cloud.fogbow.common.models.OpenStackV3User;
 import cloud.fogbow.ras.constants.Messages;
 import cloud.fogbow.ras.constants.SystemConstants;
-import cloud.fogbow.ras.core.plugins.interoperability.openstack.network.v2.GetSecurityGroupsResponse;
+import cloud.fogbow.ras.core.plugins.interoperability.openstack.util.serializables.v2.GetSecurityGroupsResponse;
+import cloud.fogbow.ras.core.plugins.interoperability.openstack.util.serializables.v2.GetSecurityGroupsResponse.SecurityGroup;
 import org.apache.log4j.Logger;
 import org.json.JSONException;
 
@@ -33,7 +34,7 @@ public class OpenStackCloudUtils {
     public static String getSecurityGroupIdFromGetResponse(String json) throws InternalServerErrorException {
         String securityGroupId = null;
         try {
-            GetSecurityGroupsResponse.SecurityGroup securityGroup = GetSecurityGroupsResponse.fromJson(json).getSecurityGroups().iterator().next();
+            SecurityGroup securityGroup = GetSecurityGroupsResponse.fromJson(json).getSecurityGroups().iterator().next();
             securityGroupId = securityGroup.getId();
         } catch (JSONException e) {
             LOGGER.error(String.format(Messages.Log.UNABLE_TO_RETRIEVE_NETWORK_ID_S, json), e);
